% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animap_vector_cl.R
\name{animap_vector_cl}
\alias{animap_vector_cl}
\title{animate maps of larvae control}
\usage{
animap_vector_cl(path, locality, path_loc, vel, dir, name, x_leg, y_leg)
}
\arguments{
\item{path}{is the directory of the larvae control file.}

\item{locality}{is the locality target.}

\item{path_loc}{is the directory of shepefile dataset for limit locality.}

\item{vel}{is the delay time between images. See also \link[tmap]{tmap_animation} and  \link[tamp]{fps}}

\item{dir}{is the directory where the animation will be saved.}

\item{name}{is the name of the gif file.}

\item{x_leg}{is the x position of legend.}

\item{y_leg}{is the y position of legend.}
}
\value{
a gif file of animation.
}
\description{
this function created the animated map for larvae control.
}
\details{
\link[INLA]{inla}.
}
\examples{
1+1
}
\seealso{
\link[tmap]{tmap::tmap_animation}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
