% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks_surveillance.R
\name{blocks_surveillance}
\alias{blocks_surveillance}
\title{blocks_surveillance}
\usage{
blocks_surveillance(locality, cve_edo)
}
\arguments{
\item{locality}{is the locality target.}

\item{cve_edo}{is the numeric id of state.}
}
\value{
a mapview object.
}
\description{
this function generate the map of block surveillance
}
\examples{
1+1
}
