% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_eggs_risk.R
\name{cluster_eggs_risk}
\alias{cluster_eggs_risk}
\title{Cluster analysis for egg hotspot intensity and entomological risk}
\usage{
cluster_eggs_risk(x)
}
\arguments{
\item{x}{is the eggs hotspots dataset}
}
\value{
a sf object
}
\description{
Cluster analysis for egg hotspot intensity and entomological risk
}
\examples{
1+1
}
