% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords}
\alias{coords}
\title{Polygons of blocks from Mexico of the National Electoral Institute 2020}
\format{
A sf object with 2278337 rows and 19 variables:
\describe{
  \item{clave}{the ids of ovitrap.}
  \item{entidad}{key and state name.}
  \item{jurisdiccion}{key and jurisdiction name.}
  \item{municipio}{key and munucipality name.}
  \item{localidad}{key and locality name.}
  \item{sector}{key to the sector.}
  \item{manzana}{key of blocks.}
  \item{ovitrampa}{key of blocks.}
  \item{tipo_de_escuela}{type of school.}
  \item{clave_sep}{key of sep.}
  \item{posision_x}{longitude.}
  \item{posision_y}{latitude}
  \item{semana_epidemiologica}{epidemiological week.}
  \item{usuario}{user key.}
  \item{fecha_captura}{capture date}
  ...
}
}
\usage{
coords
}
\description{
A spatial dataset containing of polygon data of blocks from Mexico
}
\details{
coordinates of ovitraps 2015-08-11.
}
\keyword{datasets}
