% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs_hotblocks.R
\name{eggs_hotblocks}
\alias{eggs_hotblocks}
\title{eggs_hotblocks}
\usage{
eggs_hotblocks(path_vect, cve_edo, locality, brand, wk)
}
\arguments{
\item{path_vect}{is the directory of the ovitrampas dataset.}

\item{cve_edo}{is the id of state.}

\item{locality}{is the locality target}

\item{brand}{is the palette of color, thera are two option google o slack}

\item{wk}{is the target week.}
}
\value{
a mapview object.
}
\description{
eggs_hotblocks
}
\examples{
1+1
}
