% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs_hotspots_intensity_map.R
\name{eggs_hotspots_intensity_map}
\alias{eggs_hotspots_intensity_map}
\title{eggs_hotspots_intensity_map}
\usage{
eggs_hotspots_intensity_map(
  spde_betas,
  locality,
  cve_ent,
  palette,
  years,
  name
)
}
\arguments{
\item{spde_betas}{is the betas of spde model by year.}

\item{locality}{is the target locality.}

\item{cve_ent}{is the id of state.}

\item{palette}{is the palette name package and function. example  rcartocolor::carto_pal.}

\item{years}{is the target year.}

\item{name}{is the name of palette.}
}
\value{
a ggplot map.
}
\description{
eggs_hotspots_intensity_map
}
\details{
\link[INLA]{inla}, \link[deneggs]{eggs_hotspots_week}
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
