% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs_hotspots_year.R
\name{eggs_hotspots_year}
\alias{eggs_hotspots_year}
\title{eggs_hotspots_year}
\usage{
eggs_hotspots_year(
  cve_mpo,
  cve_edo,
  locality,
  hist_dataset,
  year,
  integration_strategy,
  aproximation_method,
  fam_distribution,
  path_vect,
  palette.viridis,
  plot,
  alpha.value,
  kvalue
)
}
\arguments{
\item{cve_mpo}{is the text id of the municipality.}

\item{cve_edo}{is the text id of the state.}

\item{locality}{locality is the locality target.}

\item{hist_dataset}{is a logical value for define the dataset, if TRUE is the ovitraps historical dataset and we neen define the year to analyze.}

\item{year}{is the year to analyze.}

\item{integration_strategy}{is the integration strategy. The options are "grid", "eb" & "ccd"}

\item{aproximation_method}{aproximation is the aproximation of the joint posterior of the marginals and hyperparameter. The options are "gaussian", "simplified.laplace" & "laplace"}

\item{fam_distribution}{is the name of the family of the distribution for modelling count data. The option can be  poisson, zeroinflatedpoisson0, zeroinflatedpoisson1, nbinomial, zeroinflatednbinomial0 and zeroinflatednbinomial1.}

\item{path_vect}{is the directory of the ovitrampas dataset.}

\item{palette.viridis}{is the palette of viridis. The options are plasma, viridis, inferno & magma.}

\item{plot}{is a logical value for the plot the mesh.}

\item{alpha.value}{alpha The significance level, also denoted as alpha or α, is the probability of rejecting the null hypothesis when it is true.}

\item{kvalue}{is the parameter for define the triagulization of delauney in the inner and the outer area in the argument max.edge in the INLA:inla.mesh.2d.}
}
\value{
a list of two object (a df of betas & a list of spde of each week)
}
\description{
this function predicts the number of eggs using a geostatistical analysis with INLA & calculates the hotspots of the eggs
}
\details{
\link[INLA]{inla}, \link[deneggs]{eggs_hotspots}
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
