% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs_map.R
\name{eggs_map}
\alias{eggs_map}
\title{eggs_map}
\usage{
eggs_map(
  path_lect,
  path_coord,
  path_shp,
  leg_title,
  loc,
  size,
  alpha,
  palette_vir,
  risk,
  weeks
)
}
\arguments{
\item{path_lect}{is the directory of ovitraps dataset.}

\item{path_coord}{is the directory of coordinates dataset.}

\item{path_shp}{is the directory of shepefile dataset.}

\item{leg_title}{is title of legend.}

\item{loc}{is the target locality.}

\item{size}{is the size of point.}

\item{alpha}{is the transparency of point.}

\item{palette_vir}{is the name of pallete of viridis.}

\item{risk}{is logical value, if is TRUE return of percentil map, else only map of eggs.}

\item{weeks}{is the week number.}
}
\value{
a ggplot.
}
\description{
The function generates an egg map of Aedes Aegypti.
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
