% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs_risk.R
\name{eggs_risk}
\alias{eggs_risk}
\title{eggs_risk}
\usage{
eggs_risk(path_vect, path_coord, weeks, locality, risk)
}
\arguments{
\item{path_vect}{is the directory of the ovitrampas dataset.}

\item{path_coord}{is the directory of the coordinates dataset.}

\item{weeks}{is the week.}

\item{locality}{is the name of locality}

\item{risk}{is logical value for define the risk or map the eggs.}
}
\value{
a mapview object
}
\description{
eggs_risk
}
\examples{
1+1
}
