% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_mesh.R
\name{fm_mesh}
\alias{fm_mesh}
\title{fm_mesh}
\usage{
fm_mesh(x, k, loc, plot = NULL)
}
\arguments{
\item{x}{is the sf object and the target variable.}

\item{k}{is the parameter for defining the delaunay triagulization.}

\item{loc}{is the locality (sf).}

\item{plot}{is a logical argument. if TRUE plot the mesh else no plot the mesh.}
}
\value{
a mesh and plot of the mesh.
}
\description{
This function create the mesh for modelling with INLA framework
}
\details{
this function use \link[INLA]{inla} & \link[fmesher]{fmesher}
}
\examples{
1+1
}
\seealso{
\link[INLA]{inla.mesh.2d} & \link[fmesher]{fm_mesh_2d_inla}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
