% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{indicators}
\alias{indicators}
\title{The function provides entomological indicators for entomological surveillance with ovitraps.}
\usage{
indicators(path_ovitraps, scale, locality, fac_axes)
}
\arguments{
\item{path_ovitraps}{is the path of the folder where the data is located.}

\item{scale}{is the resolution of the analysis (ovitrap for house/ovitraps and block for blocks).}

\item{locality}{is the location of interest.}

\item{fac_axes}{is a numeric value to control the height of the secondary axis}
}
\value{
a ggplot object
}
\description{
The function provides entomological indicators for entomological surveillance with ovitraps.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
