% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_grid_points.R
\name{loc_grid_points}
\alias{loc_grid_points}
\title{loc_grid_points}
\usage{
loc_grid_points(sf, cell_size)
}
\arguments{
\item{sf}{the sf locality.}

\item{cell_size}{number of cell.}
}
\value{
a points
}
\description{
this function make the prediction grid for spde modelling
}
\examples{
1+1
}
