% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_eggs_hotspots.R
\name{map_eggs_hotspots}
\alias{map_eggs_hotspots}
\title{map_eggs_hotspots}
\usage{
map_eggs_hotspots(betas, locality, cve_edo, palette, name, static_map)
}
\arguments{
\item{betas}{is the betas of inla spde model.}

\item{locality}{is the target locality.}

\item{cve_edo}{is the id of state.}

\item{palette}{is the palette name package and function. example  rcartocolor::carto_pal.}

\item{name}{is the name of palette.}

\item{static_map}{is logical value (TRUE o FALSE). if the value es TRUE the map is static, else the map es interactive.}
}
\value{
a ggplot map.
}
\description{
The function generates a map per year of the number of weeks that each ovitrap was hotspots.
}
\details{
\link[INLA]{inla}, \link[deneggs]{eggs_hotspots_week}
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
