% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovitraps_read.R
\name{ovitraps_read}
\alias{ovitraps_read}
\title{Read the ovitrap dataset}
\usage{
ovitraps_read(path, current_year, year = NULL)
}
\arguments{
\item{path}{is the directory where the files are located. The files for the current year is in txt (currente_year = TRUE), the historic dataset is a zip file (currente_year = FALSE).}

\item{current_year}{is a logical value. The current_year = TRUE is for current year dataset (2020), else is for historic dataset (2011 to 2019).}

\item{year}{is a string value. When the current_year is FALSE, we need define the year. When current_year is TRUE the value for year is NULL.}
}
\value{
a dataframe.
}
\description{
This function has been designed for read the ovitrap datasets (historic and current year).
}
\details{
the current and historic ovitrap dataset is download of the homepage of \href{http://kin.insp.mx/aplicaciones/SisMV}{**Sistema Integral de Monitoreo de Vectores**}.
}
\examples{
1+1
}
\references{
xxxxx
}
\seealso{
\link[boldenr]{read_dataset_bol}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
