% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ovitraps_coordinates.R
\name{read_ovitraps_coordinates}
\alias{read_ovitraps_coordinates}
\title{read_ovitraps_coordinates}
\usage{
read_ovitraps_coordinates(path_ovitraps, path_coord, cve_edo, locality, sf)
}
\arguments{
\item{path_ovitraps}{path of ovitraps folder}

\item{path_coord}{path of coordinates folder}

\item{cve_edo}{id of state}

\item{locality}{locality target}

\item{sf}{is a logical value, if true return a sf object else a dataframe}
}
\value{
a sf o dataframe.
}
\description{
read_ovitraps_coordinates
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
