% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks_surveillance.R
\name{blocks_surveillance}
\alias{blocks_surveillance}
\title{blocks_surveillance}
\usage{
blocks_surveillance(locality, cve_edo, path_coords)
}
\arguments{
\item{path_coords}{}
}
\description{
blocks_surveillance
}
