% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords}
\alias{coords}
\title{Coordinates of the ovitraps of the Entomological Surveillance System in México}
\format{
A sf object with 208034 rows and 16 variables:
\describe{
  \item{clave}{the ids of ovitrap.}
  \item{entidad}{key and state name.}
  \item{jurisdiccion}{key and jurisdiction name.}
  \item{municipio}{key and munucipality name.}
  \item{localidad}{key and locality name.}
  \item{sector}{key to the sector.}
  \item{manzana}{key of blocks.}
  \item{ovitrampa}{key of blocks.}
  \item{tipo_de_escuela}{type of school.}
  \item{clave_sep}{key of sep.}
  \item{posision_x}{longitude.}
  \item{posision_y}{latitude}
  \item{semana_epidemiologica}{epidemiological week.}
  \item{usuario}{user key.}
  \item{fecha_captura}{capture date}
  \item{geometry}{geometry}
  ...
}
}
\usage{
coords
}
\description{
A spatial dataset the coordinates of the ovitraps
}
\details{
coordinates of ovitraps 2025-08-11.
}
\keyword{datasets}
