% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovitraps_occupancy_abundancy.R
\name{ovitraps_occupancy_abundancy}
\alias{ovitraps_occupancy_abundancy}
\title{ovitraps_occupancy_abundancy}
\usage{
ovitraps_occupancy_abundancy(path_ovitraps, scale)
}
\arguments{
\item{path_ovitraps}{is the path  of the ovitrap file.}

\item{scale}{is a string that define the scale, there are two options ovitraps & blocks}
}
\description{
ovitraps_occupancy_abundancy
}
