% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_pred_map.R
\name{spde_pred_map}
\alias{spde_pred_map}
\title{spde_pred_map}
\usage{
spde_pred_map(
  path_lect,
  locality,
  path_coord,
  path_shp,
  leg_title,
  alpha,
  plot = NULL,
  aproximation,
  integration,
  longitude,
  latitude,
  k,
  week,
  var,
  cell_size,
  palette_vir
)
}
\arguments{
\item{path_lect}{is the directory of the ovitrampas readings file.}

\item{path_coord}{is the directory of the ovitrampas coordinates file.}

\item{path_shp}{is the directory of shepefile dataset.}

\item{leg_title}{is title of legend.}

\item{plot}{is a logical argument. if TRUE plot the mesh else no plot the mesh.}

\item{longitude}{is the name of the column of the longitude in the ovitrampas dataset.}

\item{latitude}{is the name of the column of the longitude in the ovitrampas dataset.}

\item{k}{is the parameter for define the triagulization of delauney.}

\item{week}{is the week target. proactive is the current week.}

\item{var}{is the name of variable target.}

\item{cell_size}{is the parameter for define the grid of locality for prediction.}

\item{palette_vir}{is the palette of viridis. The option can be magma, plasma, inferno, and viridis.}

\item{loc}{is the locality target.}

\item{fam}{is the name of the family of the distribution for modelling count data. The option can be  poisson, zeroinflatedpoisson0, zeroinflatedpoisson1, nbinomial, zeroinflatednbinomial0 and zeroinflatednbinomial1}
}
\value{
a list with the gmap with the prediction of the number of eggs, the prediction of number of eggs, the mesh.
}
\description{
This function predicts the number of eggs in areas where it was not collected using geostatistical analysis with INLA
}
\details{
\link[INLA]{inla}.
}
\seealso{
\link[viridis]{viridis}, \link[viridis]{plasma}, \link[viridis]{inferno}, \link[viridis]{magma}

\link[INLA]{inla}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
