% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_pred_map.R
\name{spde_pred_map}
\alias{spde_pred_map}
\title{spde_pred_map}
\usage{
spde_pred_map(
  path_lect,
  locality,
  path_coord,
  cve_ent,
  leg_title,
  alpha,
  plot = NULL,
  aproximation,
  integration,
  longitude,
  latitude,
  k,
  week,
  var,
  cell_size,
  palette_vir
)
}
\arguments{
\item{path_lect}{is the directory of the ovitrampas readings file.}

\item{locality}{is the locality target.}

\item{path_coord}{is the directory of the ovitrampas coordinates file.}

\item{cve_ent}{is the text id of the state.}

\item{leg_title}{is title of legend.}

\item{alpha}{The significance level, also denoted as alpha or α, is the probability of rejecting the null hypothesis when it is true.}

\item{plot}{is a logical argument. if TRUE plot the mesh else no plot the mesh.}

\item{aproximation}{is the aproximation of the joint posterior of the marginals and hyperparameter. The options are "gaussian", "simplified.laplace" & "laplace".}

\item{integration}{is the integration strategy. The options are "grid", "eb" & "ccd".}

\item{longitude}{is the name of the column of the longitude in the ovitrampas dataset.}

\item{latitude}{is the name of the column of the longitude in the ovitrampas dataset.}

\item{k}{is the parameter for define the triagulization of delauney.}

\item{week}{is the week target. proactive is the current week.}

\item{var}{is the name of variable target.}

\item{cell_size}{is the parameter for define the grid of locality for prediction.}

\item{palette_vir}{is the palette of viridis. The option can be magma, plasma, inferno, and viridis.}

\item{path_shp}{is the directory of shepefile dataset.}
}
\value{
a list with the gmap with the prediction of the number of eggs, the prediction of number of eggs, the mesh.
}
\description{
This function predicts the number of eggs in areas where it was not collected using geostatistical analysis with INLA
}
\details{
\link[INLA]{inla}.
}
\examples{
1+ 1
}
\seealso{
\link[viridis]{viridis}, \link[viridis]{plasma}, \link[viridis]{inferno}, \link[viridis]{magma}

\link[INLA]{inla}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
